introduction of a prebuild image to avoid unnecessary download of data

+++ build_in_container.sh
@@ -1,20 +1,43 @@
 #!/bin/bash
+cd "$( dirname $(realpath $0))"
+
 BUILD=
 RUN=
 CN="matrissys"
 TAG="v1.0"
 VOL="$(pwd)/lfs_volume"
 
+BASIS_IMG="build-env-basic"
+BASIS_IMG_TAG="v1.0"
+
 function _help ()
 {
     echo "USAGE: $( basename $0 ) -a|--all|-b|--build|-r|--run <cmd>"
 }
 
+function _build_basic_image ()
+{
+    local ERR=0
+    docker images --format "{{.Repository}}:{{.Tag}}" | grep "$BASIS_IMG:$BASIS_IMG_TAG"
+    if [ $? -ne 0 ]; then
+        pushd prebuild
+        docker build -t $BASIS_IMG:$BASIS_IMG_TAG -f Dockerfile_basicImage .
+        ERR=$?
+        [ $ERR -ne 0 ] && echo "building $BASIS_IMG:$BASIS_IMG_TAG failed"
+        popd    
+    fi 
+    return $ERR
+}
+
 function _build_image ()
 {
+    local ERR=0
+    _build_basic_image
+    ERR=$?
+    [ $ERR -ne 0 ] && return $ERR 
     docker rmi $CN:$TAG 2> /dev/null       
-    docker build -t $CN:$TAG .
-    return $?
+    docker build --build-arg baseimage="$BASIS_IMG:$BASIS_IMG_TAG" -t $CN:$TAG .
+    return 0
 }
 
 function _run_container ()
